requirements.txt
requests==2.22.0
python-telegram-bot==12.0.0b1